refactor: move get_transforms to utils
